+++ src/App.js
@@ -1,7 +1,15 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
-import { AboutPage, ContactPage, GalleryPage, LandingPage, ProjectsPage } from 'pages';
+import {
+    AboutPage,
+    ContactPage,
+    GalleryPage,
+    LandingPage,
+    ProjectsPage,
+    RelatosVisualesPage,
+    RelatosVisualesDetailsPage
+} from 'pages';
 
 const App = () => {
     return (
@@ -11,6 +19,8 @@ const App = () => {
             <Route path="/contact" exact element={<ContactPage />} />
             <Route path="/gallery" exact element={<GalleryPage />} />
             <Route path="/projects" exact element={<ProjectsPage />} />
+            <Route path="/relatos-visuales" exact element={<RelatosVisualesPage />} />
+            <Route path="/relatos-visuales/:id" element={<RelatosVisualesDetailsPage />} />
         </Routes>
     );
 };