+++ src/pages/index.js
@@ -3,3 +3,5 @@ export { default as ContactPage } from 'pages/contact/Contact';
 export { default as GalleryPage } from 'pages/gallery/Gallery';
 export { default as ProjectsPage } from 'pages/projects/Projects';
 export { default as LandingPage } from 'pages/landing-page/LandingPage';
+export { default as RelatosVisualesPage } from 'pages/relatos-visuales/RelatosVisualesPage';
+export { default as RelatosVisualesDetailsPage } from 'pages/relatos-visuales/RelatosVisualesDetailsPage';